requirements.txt
Pillow>=9.0.0
pyexiv2>=2.0.0
customtkinter>=5.0.0
pyinstaller>=5.0.0
